Refactor: Extract ProjectCreated case logic to handleProjectCreatedEvent function
- Extracted the logic for ProjectCreated case to handleProjectCreatedEvent in ./handlers/handleProjectCreatedEvent.ts
- Updated the switch statement in handleEvent to call handleProjectCreatedEvent for the "ProjectCreated" event
