+++ src/indexer/allo/v1/handlers/index.ts
@@ -1,0 +1,1 @@
+export { handleProjectCreatedEvent } from "./handleProjectCreatedEvent.js";
